refactor: split lifecycle.ts into focused modules (#8259)
